Changed to LootLocker SDK
Changed our approach to database storage and retrieval from Firebase to LootLocker as it is much more intuitive to set up.

Further testing required to see the scale of the Leaderboard UI with the desired 10 names in the leaderboard.

When porting to MVP, check for the correct object references as there are test ones made (for testing purposes) - eg. TestItemCollector, originally on shuttlecock gameObject. Check carefully to prevent bugs.